Conditionally compile GC source files. Test stock build in CI. (#73)
This PR changes the build script to only include source files for the GC implementation that is selected. The PR also removes unnecessary preprocessor `#ifdef` for related files.

This PR adds a CI job to build with the stock GC. This helps us to make sure that the stock GC can build.