Save thread context before yielding for GC (#78)
This PR ports https://github.com/mmtk/mmtk-julia/pull/159 to `dev`. The difference is that this PR adds a general call to the GC interface `jl_gc_notify_thread_yield`. In this case, each GC will do what they need in the call, and the context is saved in the GC specific TLS.