Trace pinned objects as roots (#91)
Following the suggestion in https://github.com/mmtk/julia/pull/89#issuecomment-2982261402, trace recently added pinned objects as roots. When we pass those objects as 'nodes' to MMTk, MMTk does not know the slots so MMTk cannot update the reference, thus MMTk will not move those objects. This is essentially the same as pinning those objects before a GC, and unpinning after a GC.